fix: refine domain entities id in order to support uint64 instead of default uint32

+++ internal/domain/brokerage.go
@@ -7,7 +7,7 @@ Brokerage
 Represents Analytics Brokerage teams that are responsible for the stocks recommendations.
 */
 type Brokerage struct {
-	ID        uint      `json:"id,string"`
+	ID        uint64    `json:"id,string"`
 	Name      string    `json:"name"`
 	CreatedAt time.Time `json:"created_at"`
 }
