+++ internal/domain/market.go
@@ -7,7 +7,7 @@ Market
 Represents the market where the stock is traded.
 */
 type Market struct {
-	ID        uint      `json:"id,string"`
+	ID        uint64    `json:"id,string"`
 	Name      string    `json:"name"`
 	CreatedAt time.Time `json:"created_at"`
 }
